+++ CHANGELOG.md
@@ -20,6 +20,7 @@
     - The Node **FirmwareUpdateFinished** event now passes an args parameter
 
   - New Features  
+    - Added **SetRawConfigParameterValue** method to the ZWaveNode class.
     - Added **RefreshValues** method to the ZWaveNode class.
     - Added **WaitForWakeUp** method to the ZWaveNode class.
     - Added **Ping**, method to the ZwaveNode class.